requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
colorama==0.4.6
discord.py==2.4.0
Faker==25.2.0
faker-vehicle==0.2.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
instant-api-client==0.1.2
leet==0.1.2
loguru==0.7.3
multidict==6.1.0
orjson==3.10.15
pocketbase==0.14.0
pomice==2.9.2
propcache==0.2.1
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
websockets==14.2
win32_setctime==1.2.0
yarl==1.18.3
yt-dlp==2025.1.26
zxcvbn==4.4.28
